Include network interface name in config output
If we can find the network interface name corresponding to our cdc-wdm
control device from sysfs, include it at the beginning of the config
command output. This makes it easier for a script consuming our output
to configure the correct network interface.

Continue gracefully if sysfs is not accessible: we may be in a chroot
jail or otherwise sandboxed.