Exclusively lock the control device
If two tinymbim commands are run concurrently against the same control
device, one process can read the response for which the other process
is waiting, resulting in deadlock. The same problem is present and easy
to reproduce in OpenWrt umbim and uqmi.

Take an exclusive flock on the control device before issuing a request
to properly serialise access.