Move timeout behaviour from mbim tool to wwan script
Rather than hard-coding a one second timeout in mbim close, handle the
timeout in the wwan wrapper script.

Add some sensible timeout and retry logic to the open operation in the
script too, as another common modem bug is discarding commands for a few
seconds after the control device is registered.